Footer section (#17)
* Setup the nextjs project

In this change I set up a nextjs typescript project configured with tailwindcss for styling and prettier for formating

* Setup nextjs typescript project

In this change, I setup this nextjs typescript project configured with tailwindcss and prettier

* Deleting all installed files

In this cahange, I remove all installed files that was causing conflict with the main branch

* Setup the nextjs typescript project

In this change, I setup the nextjs typescript project configured with tailwindcss for styling and prettier for formating the codebase

* delete some unnecessary files

In this change, I deleted next.svg and vercel.svg which is not necessary to the project

* Created the navigation menu

In this change, I create the navigation menu for the eduleaening and made responsive to view in mobiles as well as desktop

* Fix the menu state and name convention

In this change, I made an updater function in the toggle menu function and made changes in the naming convention using the pascalcase

* Fix the menu state and name convention

In this change, I made an updater function in the toggle menu function and made changes in the naming convention using the pascalcase

* Create the Hero Into section

In this change, I create an intro scetion for the hero page

* Fis the current index

In this chnage, I remove dthe quote state and changed the currentQuote to currentIndex

* Fix the hero section image

In this change I changed the hero section image type from png to svg in which the image was blurry

* Create an offer section

In this change, I create an offer section on the home page

* feat: create what we offer section

In this change, I create an what we offer section

* made the offer section responsive

In this change, I made the offer section responsive

* create the hero category section

In this change, I create the category section in the home page

* updated the hero category section

In this change I updated the hero category section

* updated the mobile version of hero category section

In this section I update the responsivess for the hero category section

* change the view all categories button to Link

In this change, I change the view button to Link

* create the courses her section

In this section I create the courses section for the hero section

* remove the import react and improve the videos section code base

In this change I upadate and improve the codebase

* change the props type in the offers and videos

In this change, I change the props type for the courseCard and the offerIcon

* change the icon type to ReactElement

In this change, I changed the icon type to ReactElement

* change the icon type in the OfferIconProps

In this change I change the icon type in the OfferIcon Props

* Remove the import 'React'

In this change, I removed the import 'React' from the offer page in the home page

* formatted all files

In this change, I formatted the files using prettier

* set up the more section

In this change, I set up the more section on the home page

* Setup the footer component

In this change I created the logo component for the foooter

* set up the footer component

In this change, I set up the footer components by adding the category component, useful links component and the quick links component

* Implement Footer section

Added a new footer component that includes a logo, category links, useful links, and a subscribe form. The logo component uses React Icons to display a book icon and includes social media links. The category and useful links components display lists of links with custom styling. The subscribe form allows users to input their email address and submit it to receive newsletters

* made heading changes

In this change the heading in the footer page was changed

* reset up the footer and navbar component

In this change, the footer and navbar component was reset up

---------

Co-authored-by: Motunrayo Adeneye <motunrayodeneye@gmail.comgit config --global user.name Motunrayo>